pyproject.toml
[project]
name = "capamedia-cli"
version = "0.1.0"
description = "CLI multi-harness para la migración de servicios legacy (IIB/WAS/ORQ) a Java 21 + Spring Boot hexagonal OLA1 en Banco Pichincha"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Banco Pichincha - Capa Media Team" }]
keywords = [
    "migration",
    "legacy",
    "iib",
    "was",
    "spring-boot",
    "claude",
    "copilot",
    "cursor",
    "windsurf",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "jsonschema>=4.21.0",
    "tomli>=2.0.1",
    "tomli-w>=1.0.0",
    "pyperclip>=1.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[project.scripts]
capamedia = "capamedia_cli.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/capamedia_cli"]

[tool.hatch.build.targets.wheel.shared-data]
"src/capamedia_cli/data" = "capamedia_cli/data"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
